Update patch set 2
Patch Set 2:

Leszek, I find this frustrating. Why was it not possible to merge this patch and do a follow up? Why did I got no response for three days? I was not in the office, but from what I see in my inbox it appears the entire team tried to understand and solve the issue that was already solved by this patch more than a week ago. Could it be that the -1 here wasted multiple days of combined team effort?

Patch-set: 2
Reviewer: Thiemo Kreuz (WMDE) <1453@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0