Update patch set 2
Patch Set 2:

> Why was it not possible to merge this patch and do a follow up? Why did I got no response for three days?

As I said in the comment, I don't think this patch was doing a right thing. Sure, I could have go on with discussing further why I disagreed etc, but I found submitting the alternative patch doing what I think was "right" was more constructive.
Regarding the "why not do a follow up" part: I didn't think merging something that does "wrong" (note quote marks) thing and put a patch "fixing" it on top of it was the best thing to do, when it was possible to "fix" it "all" in a single patch. Doing the latter and creating a follow up seemed the same amount of work, so why not do the "right thing" right away?

> from what I see in my inbox it appears the entire team tried to understand and solve the issue that was already solved by this patch more than a week ago. Could it be that the -1 here wasted multiple days of combined team effort?

To be honest, I didn't get this part. Is the last sentence a question you would like me to answer?

Patch-set: 2