[DNM][EXPERIMENT] Require *-term permissions to edit entity terms using wbeditentity
This is a rough experiment to see what it takes to make wbeditentity
be more accurate about permission its changes require.

T170673 suggests other way to solve this issue.

This makes EditEntity get information required to decide on what permissions
are required from EntityChangeOpProvider, which in turn gets this from
entity-level ChangeOpDeserializers (using PermissionAwareChangeOpDeserializer
interface).

Possible advantage compared to what T170673 suggests: as the issue
seems to be only relevant to wbeditentity, ChangeOp interface is not
"polluted" with permission-related stuff, and stays simple.

On the flip side: solution here is a bit messy.

Also adds permission-related tests for creating an entity using
setX API actions which have not been there before. Also adjusts
WikiPageEntityStorePermissionChecker to handle the case of creating
the entity when peforming ACTION_EDIT_TERMS.

TODO:
 - resolve TODOs (e.g. what about wbeditentity's "clear" param?)
 - add tests for missing new bits
 - this is probably too much change for a single patch!

Change-Id: I294f7bcfbd0e10b6a6ad82c80ddd8a4ed5eb637e