Trim earlier in SpecialGoToLinkedPage and test it
The trimming of the individual values from the sites array was always
done, but very late. Now it's all done as early as possible.

I also replaced an other trim() with the one from the stringNormalizer.

Plus I changed the test and added a space after a comma to actually test
this.

Bug: T166473
Change-Id: I70d9e1e35e2c04988af8eb32c3ec2975eddc3662