Update patch set 3
Patch Set 3: Code-Review-1

(2 comments)

I'm sorry, but I can't tell what's going on here. The file mentioned in the commit message is not touched. I find the explanation why this is needed not sufficient. There are no tests, and no test is changed. As this patch is now this is just dead code that does not have any effect, and should not be merged because of this. Please create a ticket or upload a more complete patch that makes the issue this change aims to solve more obvious.

Patch-set: 3
Label: Code-Review=-1