Update patch set 2
Patch Set 2:

I'm afraid I don't understand what you mean. https://github.com/wmde/WikibaseDataModel/pull/718 is unrelated. That change was about PHP's internal serialize() feature. This patch here is about our JSON serialization. It still contains the entity type and numeric ID, see https://github.com/wmde/WikibaseDataModel/blob/master/src/Entity/EntityIdValue.php#L116. All this is just fine as it is now, with the Int32… interface in place. So we will most probably never remove these array elements.

The question for this patch here is: Is it worth it? The only effect of this patch is that it minimizes the database footprint a little bit. Is this an optimization we care about?

This will not make dumps smaller.

Patch-set: 2