Update patch set 1
Patch Set 1: Code-Review+1

This was introduced in 2015 with I632833a to make the new implementation compatible with what the old serializer did. The old one skipped empty elements (on the first level of the entity serialization). We changed this everywhere, except here, for no reason. We just forgot.

Open question is, how critical is it to announce this? Gadgets and user scripts might behave a little weird in edge cases. I don't think anything can break. The worst case I can think of is a gadget starting to show something like "here is a list of labels", but there is nothing displayed.

I will add a note to the weekly. Do others think this needs more announcement?

Patch-set: 1