Update patch set 1
Patch Set 1:

I do not consider this a breaking change (adding keys is explicitly defined not to be a breaking change). It would still be nice to announce it, but we don't need to do that in advance.

Patch-set: 1
Reviewer: Daniel Kinzler <128@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=0