Update patch set 3
Patch Set 3: Code-Review-1

This obviously breaks a lot of stuff. This should be easy to fix. The question is more like: is this worth it? This would change the internally stored JSON blobs. But these are not meant to be consumed by anybody. The contents of the revision table are explicitly not stable. The contrary. You might find very, very old serializations in there that are entirely not what you expect.

I'm not objecting this change. I believe we can do this, but I want us to understand the impact better.

Patch-set: 3
Reviewer: Thiemo Kreuz (WMDE) <1453@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=-1