Update patch set 3
Patch Set 3:

> I would like the JSON serialization returned by the 'revisions' API to be more predictable

The JSON returned by the revisions API is not stable. It's subject to changes without notice. It is useful only as an opaque blob that can be exported and imported. 

So, making it "look a bit nicer" will only misguide more people into using that data. If you need JSON of a specific revision, use Special:EntityData

Patch-set: 3