Update patch set 6
Patch Set 6:

Okay, I see the separate changes now. I forgot about the fact that we're not just consolidating the use cases, but also switching page cache invalidation from in-process to job-queue. That is indeed worth keeping in a separate change. However, if the intend is to merge and test these separately as workable point in time (which I support), then the three commits should be re-ordered as follows:

Current:

1. Use RefreshLinksJob instead of Title:invalidateCache.
2. Use multi-page RefreshLinksJob.
3. Remove redundant abstraction.


Proposed:

1. Use multi-page RefreshLinksJob.
2. Use RefreshLinksJob instead of Title:invalidateCache.
3. Remove redundant abstraction.

Reason: Using jobs for each page invalidation seems like a bad idea and not something we should assume to safely "revert back to" if needed. Rather, that should happen first, and then its usage widened.

Patch-set: 6