Update patch set 1
Patch Set 1:

(1 comment)

Thanks for jumping in and refactoring this. Any help is very welcome. The only little issue I have with the process here is that this was merged only one hour later. Nobody else had time to review this. Please try to wait at least 24 hours so we have a chance to see such stuff. Not necessarily because I want to find mistakes, ;-) but because I want to learn.

Patch-set: 1