Update patch set 6
Patch Set 6: Code-Review-1

(6 comments)

The -1 is only for the commit message. All my other comments are just notes and suggestions on code style and such. Feel free to disagree. The code looks fine otherwise. All it does, after all, is removing a mode. And that's my problem: The commit message does not explain why, and how the now used mode is equivalent to the removed one.

Patch-set: 6
Label: Code-Review=-1