Update patch set 3
Patch Set 3:

(1 comment)

@Ricordisamoa I'm still not quite sure I fully understand the intent of this, see my inline comment.

I agree that it would be very nice if the revision API would return data in a way that is consistent with how data is returned from the API. However, I'm not sure we can guarantee this. I think the revision API uses the same mechanism that is used for storing entities in the database. For storage, we want to omit any derived data (such as property types in statements). So there will probably always be differences.

In the external "canonical" representation, all defined keys should always be given, so empty arrays should be included.

Patch-set: 3