Update patch set 3
Patch Set 3:

(1 comment)

What I care about is a way to retrieve old versions of entities in a way that is as consistent as possible with wbgetentities. Preferably, I'd also like to be able to retrieve multiple revisions with a single request.

The "canonical" wbgetentities representation always serializes the "labels", "descriptions", "aliases", "claims" and "sitelinks" as JSON objects, even when they are empty, while Special:EntityData serializes them as JSON arrays when they are empty.

Patch-set: 3