Allow "unitStorage" configuration without referencing class names
The fact that *configuration* encodes a full qualified class name as a
string is always very suprising. This just broke the train, see T171107.
This should be avoided if not really, really necessary, and indeed, it is
not necessary here. All the code is in Wikibase anyway: the code that
constructs a UnitStore, as well as the two classes that can be used.

What the old configuration (with the full qualified class name) allows is
to introduce a new UnitStorage subclass that is *not* in the Wikibase.git
code base. I doubt this will ever be needed. It appears to be over
engeneered.

This patch keeps a backwards-compatibility mode as long as needed and
therefore can be merged any time.

Bug: T171107
Change-Id: I3b39415c0a9ca32307430e30dc7100be970d9339