Add name of the parameter to the "param-missing" API error
I'm only touching cases where a single parameter is missing. This patch
leaves a few usages untouched where this message is used for more than
one parameter. I leave this for a later patch to fix (also I0d2a471
already fixed most of them). Note that this patch here does not introduce
a new issue. The unlocalized text descriptions of all "param-missing" are
currently entirely hidden because the message *exists* (even if untranslated).

Proof:
https://www.wikidata.org/wiki/Special:ApiSandbox#action=wbgetentities&sites=dewiki

The unlocalized description for this error is "Must request one site, one
title, or an equal number of sites and titles", but this string does not
surface any more.

Change-Id: I3900f0b8d158aeeeb74b9caf713ce9ac6f5a4ad2