Add composer/installers dependency
This is what I understand: This dependency is recommended for all
repositories that are marked as "type": "mediawiki-extension". Without
this, Composer will install the component in "vendor". With this,
Composer understands that repositories marked as "mediawiki-extension"
should go to the "extensions" directory.

I might misunderstand something here. Please tell me.

Change-Id: Ib2a5dc1b2bdf44552e4a8b8e4f59adf9b16e8d91