Update patch set 28
Patch Set 28:

> The issue i mentioned in my comment on PS26 remains unresolved: should we continue to use "O" to track statement usage as "other", or should we use "C" with no modifier?

IMO we should continue using "O" for now, especially given how this is going to be used for evaluation initially (in a way which allows us to revert this anytime w/o any problems).
In a follow up, we probably want to make it possible to only write "C" usages and make "O" not include Statements, but that's for another change, as it can be done at any point in time, I think.

Patch-set: 28