Update patch set 1
Patch Set 1:

(1 comment)

> I really don't think this can be a reason for a -1. Sure you "don't like" it. You wrote the original code.

This is not the reason for -1 (and I resent the implication btw, and ask to not make those in the future, please), and I explained what the reason is - the code hardcodes the fact that config can be loaded only from file (if the FIXME is implemented), and only from files of certain extensions, and would require modification of this particular code if we ever wanted to load configuration in any other way. This is an inflexible design. Even worse, it assumes every file with extension not .json is a CSV file. Also doesn't seem to be a good idea to me. 

> There is not more "hard coded" than before

With FIXME at least, the is. Before, the class + config args decided how to read the data, so you could use any class that uses any means to get the config. With the patch, config has to be stored in a file, and only two specific file formats are ever supported by the code, and every file that isn't .json is assumed to be CSV (which btw would include ".JSON" on a case-insensitive filesystem). If you want any more, you'd need to edit this specific code. 

>  the code now contains a map that maps file extensions to class names. 

Which is hardcoded in the code. I would prefer more flexible approach. 

> using the full qualified class name (something I strongly believe should be avoided at all costs

Why? Many configs use classnames in Mediawiki. Check the mediawiki-config repo - many class names are in configs. What is the problem with it?

> This patch does not block anybody from introducing other methods,

I'm not sure I understand this. If you mean to keep both the getObjectFromSpec and filename configs, please rewrite it so it doesn't rely on exceptions (e.g. checking whether config is string or array, or any other method you like) for branching, and then it will be acceptable as the flexibility does not suffer. However, that contradicts you FIXME in the code. However, if you plan to remove getObjectFromSpec, as per FIXME, then my objection is that it reduces flexibility and there's no reason to do that.

>  Therefor I ask other reviewers to ignore the -1 

My opinion, as a person who designed and implemented this particular subsystem, that this patch is not ready for merge yet, it needs some improvements. This is the meaning of -1 - patch needs improvements, and I described which improvements I'd like to have. If you think my opinion on this matter has no value, well...

Patch-set: 1