dispatchChanges: Release client lock if an Exception occurs
Holding the lock for "ever" is very bad and can lead to short
hiccups breaking the entire dispatch process for a fair bit.

Writing a proper test for this is probably very hard, thus I
didn't both for now (please note that the rest of the exception
handling code is untested as well, AFAICT).

Change-Id: I12ffff189bcc0421e5b9c855bf4e0c0c7f8ec55b