Use proper DB name for local wiki in wb_changes_dispatch table.
The database name was previously recorded as "0" for the local database
if the list of client databases represented the local db name as `false`,
as is common in MediaWiki.

Note that this change should not have any effect on the live site:

Firstly, the changed code will only affect what value will be used
when inserting new rows into changedispatch. This only happens when
new wikis are added to the client list. For wikidata.org, the local
(repo) wiki is already present in the table.

Secondly, this is only relevant if the local wiki is represented as
`false` in the list of clients. This is not the case for the live
site.

Thirdly, the value of the chd_db field is only used for logging and
reporting. It has no impact on behavior.

Change-Id: I4195e08b2345880f8380f76c329544d603c804d1