Update patch set 1
Patch Set 1:

This is an edge case related to how the setting is documented. The documentation says "mapping global site IDs to logical database names". 
The logical database name for the local database is, by convention, false.

But I agree that it may be more useful to fail when encountering a non-string here.

Patch-set: 1