Update patch set 5
Patch Set 5:

(1 comment)

Previously, it was possible to explicitly disable conflict checks, in addition to using the current revision as the base revision. That distinction is pointless, the behavior ends up being the same.

Previouly, there was also special case handling for disabling the detecting of "late" conflicts. This too is redundant.

This change should not change the effective behavior at all. It just removes pointless options and special case code.

Patch-set: 5