Update patch set 1
Patch Set 1:

I could agree that the way the configuration is done currently is not beautiful but at least it is clear what class will be used, and it if one wanted to use custom unit storage, it is possible to do in a simple way (not sure if such possibility is needed, though).

What I can see with the changes suggested:
 - what unit storage implementation used is no longer clear, it is hidden somewhere deep in code
 - it is not really possible to use a custom unit storage. It is either JSON file or CSV file, period. Maybe it is all what is wanted and it is good to keep like this. I can see that the old code was more flexible.

Therefore, I don't really like the change either. But I'll be happy to be convinced that arguments above are not really relevant here.

Patch-set: 1
Reviewer: WMDE-leszek <2626@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0