Update patch set 5
Patch Set 5:

Regarding the `EditEntity` class:
I see, most of the changed lines are touching comments/docs. 
I see it as a big fat warning - code doesn't speak for itself.

Also, I've spent half of the day playing with this code trying to understand what is happening and now can say that this code is unmaintainable.
I believe that what needs to be done is refactor the tests and the class first so developers unfamiliar with this code could understand what is happening.

Some of the issues I've noticed:
* Some methods are public ONLY for the sake of testing
* Getters change the state
* Getters being called just to change the state
* State of execution is stored in properties instead of local variables
* Tests rely on internal implementation, specifically on the order in which getters are called
* Tests are hard to read and they don't give any insight about what is being tested

Regarding mentioned above, I would suggest refactor this functionality first and then continue with the changes. If we delay that it will NEVER happen and situation will be becoming worse and worse over time.

Patch-set: 5
Reviewer: Aleksey Bekh-Ivanov (WMDE) <4089@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0