Update patch set 4
Patch Set 4: Code-Review-1

I just realized that my changes to the unit tests were pointless.

We only use UNION on MySQL, but since self-UNION doesn't work on temp tables, we can't do that in unit tests. 

So the new logic has to be verified manually on MySQL. I'll do that.

Patch-set: 4