Remove create action from EntityPermissionChecker interface
No code explicitly asks for create action permissions only.
It has only been a MediaWiki-permission-specific detail that
when trying to edit an entity (or its terms, etc) that does
not happen to exist yet, some additional permission might
be also required.

As such, ACTION_CREATE does not belong to the public interface,
and this makes it an internal detail of WikiPageEntityStorePermissionChecker.
WikiPageEntityStorePermissionChecker::ACTION_MW_CREATE could
possibly become a private constant once the code moves up to PHP 7.

Change-Id: I3b5e1dccd9388474981ab0e1a004f16adc940869