Update patch set 2
Patch Set 2:

I did the rebase.

You are absolutely right about the size of this patch. I would love to continue working on this, but will not have that much time the next days. But I did a quick fulltext search and found the three most used classes. Perfect candidates for separate patches:

* 23 times "\TermIndex"
* 27 times "\EntityIdLookup"
* 43 times "\EntityRevision"

Feel free to make a chain of patches for these, and rebase this one on top of the chain.

Patch-set: 2