Use the right entity serializer in EntityRevisionLookup provided by Repository Service Container
Use a serializer for configured entity types when Repo functionality
is enabled, and a ForbiddenSerializer (to be 100% secure) when
only Client component is used.

Note that with this change WikibaseClient and/or WikibaseRepo
might be using two instances of the entity serializerc,
one used by services provided by WikibaseServices container,
and another one provided by top-level factories themselves.

This is temporary and will be fixed by having WikibaseServices
provide a single instance of this service.

Change-Id: Ia98a4dcf5a5d534a19909eb115307d1ed63d0cbc