Update patch set 2
Patch Set 2: Code-Review-1

(1 comment)

I ordered the entries by Phabricator ticket number in PS2. No other change was done.

I had a look at the discussion in T151186 but I can not see a clear agreement there. I also do have a minor concern because this is the first time we are introducing an upper case language code. I'm setting a temporary -1 because of these two questions that, in my opinion, need to be discussed first.

Patch-set: 2
Label: Code-Review=-1