Update patch set 1
Patch Set 1:

Yes, this is one of the reasons this file exists.

The other one is so that the real name of the person is shown, instead of an alias. In my case, for example, I added myself to get rid of the "(WMDE)" suffix. I added more names for the same reason.

Patch-set: 1