Update patch set 5
Patch Set 5:

Daniel was so kind to quickly explain the issue and this patch to me. I hope to be able to review it soon. If anybody else can help, please do so! Even a basic +1 review without you understanding the impact of this change is very much helpful! Thanks.

Patch-set: 5
Label: Verified=0