Avoid fetching terms more than once in BufferingTermLookup::prefetchTerms
Currently we often call out to BufferingTermLookup::prefetchTerms
via DispatchingTermBuffer::getTermsOfType with the same arguments
during client page parses (of pages with a lot of Wikibase content
on). This causes a lot of SELECTs on the wb_terms table to happen
(set a break point in TermSqlIndex::fetchTerms to see all of these).

This should reduce the amount of terms table queries in production
by a fair amount (I don't have any numbers, though).

Change-Id: I71687d7f929d0dd99e76f4d667637c1234dec336