Drop the "+1" in both "(5+1 references/sitelinks)" messages
We think this was overengeneered and want to remove this additional
blue number for a while now. Instead, the counter always shows the
total number now, no matter what item was already saved and what was
not.

Note that this counter is broken anyway. It gets confused when you
start removing and adding multiple items the same time.

Change-Id: Ieed4c65ff7000aab09e74aa152fd0a7968af3e33