Update patch set 9
Patch Set 9: Code-Review-1

(1 comment)

Thanks a lot for this patch! We wanted to remove this confusing functionality for a while now. Unfortunately the change as it is now leaves the UI in a state that can become broken and unusable in certain situations. For example, it's not possible to remove certain sitelinks. It may even end in a situation where you can't add an other sitelink.

I tried to fix this in my follow up patch, but I am not sure if what I did there is enough.

Patch-set: 9
Reviewer: Thiemo Kreuz (WMDE) <1453@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=-1