Remove redundant CalendarHint from TimeInput expert
This code was meant to show the calendar model, along with a button
to switch between Julian and Gregorian. All this is entirely redundant
for a while now, because:

* We are showing the calendar model as a superscript text in the preview.
* The dropdown to manually change the calendar model is not hidden in a
  collapsed section any more.
* The link to toggle the calendar model is partially broken anyways (it
  only updates the dropdown).

Right now the calendar model is shown 3 (!) times, and there are 2 UI
elements to change it.

This is already approved by Jan (UX).

Bug: T172916
Change-Id: I55a3f09f8ac7bc6329e75ab5d738130c4984b61b