Update patch set 3
Patch Set 3:

(1 comment)

> who is "we"?

Sorry, but I don't want to get into a mode that might be misunderstood as blaming individuals. When I say "we" this generally refers to the Wikidata team as a whole. I'm using this word whenever I do have the feeling that all people I talked to about a specific topic agree with it. Since I'm not talking to every single developer about every change (this would be counterproductive in my opinion) I might miss somebodies opinion. So if you disagree with this change, please say so.

Patch-set: 3