Update patch set 5
Patch Set 5:

When the entity revision is stored in cache, the whole object is "serialized", including the class name.
Note that cache is not necessarily a database in a sense of a SQL database etc (although it could possibly). I believe it is also not the case (the DB part) for wikidata production setup.

Patch-set: 5