Add badges suppport to NewItem::withSiteLink and andSiteLink
I wrote this because I wanted to use it in the client directory, but
then realized I should not introduce a dependency between repo and
client. I know we will need this feature for sure so I'm submitting it
now even if it is currently unused.

No tests because this is a test-only class that basically tests itself
when it is used.

Change-Id: I056b21d85231a8e05a2b582267c2bbc6e7a63fc4