Update patch set 4
Patch Set 4:

This removes a *ton* of test code, and I'm not so sure any more we really, really do not want this feature any more. If we are sure we never want it again then yes, this is an awesome patch! 400+ lines of code gone, this is always cool. ;-)

In PS3 I removed a test I think is not needed any more. PS4 is nothing but a rebase.

Patch-set: 4
Reviewer: Thiemo Kreuz (WMDE) <1453@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0