Update patch set 2
Patch Set 2: Code-Review-1

(3 comments)

Even though the methods are trivial, I would like to see tests for the changes in EntityAccessor, Scribunto_LuaWikibaseEntityLibrary and WikibaseLuaEntityBindings.

An integration test for this would also be nice (that would probably be a separate test case, only for this, somewhat similar to Scribunto_LuaWikibaseLibraryInProcessEntityCacheTest). I can also write this, if you don't want to.

Patch-set: 2
Label: Code-Review=-1