Update patch set 10
Patch Set 10:

Yes, and I worked hard to avoid these duplications. The original messages accept $1 parameters because I introduced them. You should be able to see this in the diff: The $1 parameters I added are reverted in this patch now, and the duplicate …-1 messages with $1 parameters are introduced instead. This is not only not needed any more, it will also break all usages of the existing messages.

Patch-set: 10