Update patch set 1
Patch Set 1:

I think the real issue is that the behaviour of SqlEntityInfoBuilder is kind of non-deterministic. IMO it should not matter if you called removeMissing before or after the collectTerms. At least the interface does not mention this. I am trying to make some code related to this issue, stay tuned!

Patch-set: 1