Stop showing meaningless version numbers on Special:Version
Special:Version is fine with this. It falls back to a dash. The git
branch and date are still shown, if available.

Go to https://www.wikidata.org/wiki/Special:Version and be amazed how
many other extensions, skins and components do the same. Especially the
most recently touched extensions all show a dash instead of a version
number.

Change-Id: I803c328091416d9be05cc1c8d63d8d97ab3b9f7e