Update repo/client/lib/view version numbers to 1.0.0
I suggest to do this for the release before WikidataCon, and have the
versions updated as a birthday present, after they got stuck at 0.5 for
multiple years.

Note that this patch alone does not mean anything. It does not mean we
must maintain stable releases. It's really just about what you see at
[[Special:Version]].

Change-Id: I803c328091416d9be05cc1c8d63d8d97ab3b9f7e