Collect terms for redirects no matter if EntityInfoBuilder::removeMissing was called
Behaviour of SqlEntityInfoBuilder::collectTerms (and of the class in broader sense
to some extent as well) is not entirely deterministic, in a sense
it returns different result e.g. depending whether
removeMissing was called or not.
This does not seem a right thing. This patch changes the way terms
are collected, so that terms of redirects should appear correctly
on pages.

SqlEntityInfoBuilder should probably be refactored, and EntityInfoBuilder
interface reconsidered.

Bug: T96553

Change-Id: Ia128dec98ccf45dbb7835e8736fe5bab2ef15cff