Update patch set 1
Patch Set 1:

I have not looked at this patch yet, but please note that the behavior as described was indeed intentional!

The idea is to provide a set if IDs, allow it to be manipulated, and then load additional information for the things in the set. This is a stateful builder, instances are not intended to be re-used! 

I'm all for refactoring if it helps. I'm not sure it can be made really pretty, though. This is a batch processing interface designed with performance in mind. I decided against abstraction and nicety in favor of speed in several places.

Patch-set: 1