Update patch set 1
Patch Set 1:

The motivation is, that this file technically is a json file. 

For other linter-configs we use file-extensions to hint to the type of content and I would say it should be best practice. 

For stylelint this was not possible until some recent version we now use, because before that a .stylelintrc.json was not recognized as config file.

Patch-set: 1