Update patch set 1
Patch Set 1:

>  Doesn't that solve whatever problem this change is supposed to address? I guess I don't really understand the problem.

So the problem is that EntityParserOutputGenerator::getEntityInfo calls the following methods of EntityInfoBuilder in the given order:
 - resolveRedirects
 - removeMissing
 - collectTerms

and the terms of the redirect "source" are not collected with the current code (i.e. w/o this change).
Could be that the "right" solution to this problem would be to change the order of calls of collectTerms and removeMissing (as in Id9ce769713a3e2438b7513d98263534609867bd6) but it seems as not a very nice solution to me.

> I'm not sure I understand the intent of this change. removeMissing() has a parameter for controlling hwether redirects will be removed or not.

the default value of that parameter, i.e. 'keep-redirects' is used and in my understanding it does not work in this case, given T96553 (which I can reproduce locally).

> The idea is to provide a set if IDs, allow it to be manipulated, and then load additional information for the things in the set. This is a stateful builder, instances are not intended to be re-used! 

I could get the point of having the stateful builder but isn't EntityParserOutputGenerator::getEntityInfo the only place EntityInfoBuilder/EntityInfoBuilderFactory are used? In this case it seems to me it is clear what set is wanted, it is clear we want redirects and want drop IDs not used on the page (including keeping the redirects - param does not seem needed given removeMissing is only used in the single place). Or do I miss something (possibly other extension also using this I guess?)

Patch-set: 1
Reviewer: WMDE-leszek <2626@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0