Update patch set 1
Patch Set 1:

This is on the .14 branch which is currently deployed.
https://github.com/wikimedia/mediawiki-extensions-Wikidata/blob/wmf/1.30.0-wmf.14/extensions/Wikibase/lib/WikibaseLib.php#L58
It might be an idea to wait until .15 for us to merge this one though incase we have any sort of unexpected roll backs and jump forwards!

Patch-set: 1
Reviewer: Addshore <106@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=0