Remove data-access dependency on WikibaseClient::getStringNormalizer
This makes GenericServices provide a StringNormalizer instance
to the data access service instantiators.

Note that with this change WikibaseClient and/or WikibaseRepo
might be using two instances of LanguageFallbackChainFactory,
one used by services provided by WikibaseServices container,
and another one provided by top-level factories themselves.

This is temporary and will be fixed by having WikibaseServices
provide one single instance of this service.

With this changes, it is no longer needed to pass WikibaseClient
instance to the callbacks instantiating services defined in
the RepositoryServiceWiring.
Note: this does not yet remove entirely dependency of data-access
on the Client component.

Change-Id: Ia2ff94c437f5f5e5cde2d3c948f07661798eb839