Do not pass the WikibaseClient instance to PerRepositoryServiceContainer
PerRepositoryServiceContainer gets the array of callbacks instantiating
entity deserializers in the constructor, instead of WikibaseClient
being passed.

PerRepositoryServiceContainerFactory gets the instance of
EntityTypeDefinitions that is then used the get the array
of deserializer "factory" callbacks.

Note this does not yet entirely remove the dependency of data access
component on the Client component.

Change-Id: I6338881ae4140894b2cb869f0216431dec32c87b