Add snak hashes to entity HTML
The snakview tempate is updated to include a class for the snak hash,
similar to the classes for statement GUIDs and reference hashes. The PHP
SnakHtmlGenerator and the JS snakview both learn to populate and update
this class.

TODO: This requires new versions of the JavaScript data model and
serialization components (not yet released).

Bug: T171725
Change-Id: Ic74cb1461b6b794df56c1590c4d2bb99393a865d