[DNM][POC] Demonstrate how possibly API class could allow handling "hierarchical entity IDs".
This just a very basic draft to demonstrate a possible way
how Wikibase API classes could handle
The idea is to NOT extend the Wikibase DataModel library with the
notion of hierarchical IDs (see T165328 for reference) but to treat
the fact the input/output contains both the "parent" entity ID
and the "subentity" ID as a input/presentation detail. Instead
of making "hierarchical" IDs the actual EntityIDs, the idea is
to make it possible for API (and other relevant) code to handle
such input (if the extension like WikibaseLexeme registers possible
services), and to present it back to the user in a desired way.

In other ways, the reasoning here is to not include anything
hierarchical-ID-specific in the core Wikibase code, but make it
possible for extensions to add code making API etc able to handle
such input.

See Ia9938e328d587c3e98e82abff453c2ceada8d2fc  in WikibaseLexeme
for a rough sketch of a possible code make it possible
to get Lexeme Form/Sense data using Wikibase's generic
wbgetentities API action.

Bug: T172192
Change-Id: I6a8b45ee4f185382672c73f09d8d8a8c03a5690a