Simplify methods providing entity serializer and serializer factory in top-level factories
Those methods have only been used with two possible option combinations,
basically being either include snak hashes in the serialization, or
don't. Generic methods that were not needed to be that generic, are
split to two methods with clear contract.

As it seeems WikibaseClient was only used to provide serializer, or
factory that only use the "compact" mode (no snak hashes), so the
other counterparts are not provided this class.

This also simplifies GenericServices::getEntitySerializer in a similar way.

Change-Id: Ia143563ed10aea81f72cb2888ed838d845e92001