Update patch set 1
Patch Set 1:

(1 comment)

Can we please proceed with this one? This is needed in a series of tests that need SiteLinks with badges. These tests can not be simplified right now because the NewItem class does not support badges, but does support SiteLinks. This is incomplete and just awkward…

Patch-set: 1
Label: Verified=0