Update patch set 3
Patch Set 3:

"Merge bubble". ;-) Never heard this term. But I understand what you mean. My personal opinion on this is: If a patch really was created 3 weeks ago, and there is no need to do a rebase, it's factually correct to have exactly this "merge bubble". An artificial rebase creates more confusion, in my opinion.

Patch-set: 3