Update patch set 2
Patch Set 2:

Thanks Legoktm for the early feedback, I really appreciate it. To be honest, I meant to to submit this as a draft but failed. I am not sure if this the direction we want to take either. I just decided to give this a try while doing some other stuff on Friday afternoon, and was curious if it would simply work on Jenkins.

Actually, it does seem more or less fine if those 4-5 JS libs are going to be kept separate, at least from the perspective of people working in Wikibase.git. But any way, let me open a subtask of Adam's mega ticket to have the discussion on what to do with those tiny JS libs there. I'll try to get some WMDE people who actually know well what those libs stand for have some word there.

Patch-set: 2