Update patch set 1
Patch Set 1:

I don't understand what your comments are meant to prove, and why you want to change this code base in the way you are proposing in this patch. What I see is a patch that touches +/- 2000 lines of code for no real reason other than a tab character. And this additional tab character doesn't even make this code easier to read. In some cases it even does the opposite.

So I ask you to update the version numbers you want to update (and I'm actually very thankful for that) without touching 2000 unrelated lines of code the same time.

Patch-set: 1