Update patch set 6
Patch Set 6: Code-Review-1

(7 comments)

This patch follows the approach of using the same RDF predicate for the  expanded URI and the literal string. That's *bad*. We want to duplicate all the relevant predicates (there is four of them for each property, I think), and use the ...-uri one for the full URI, if it exists.

Or so I think. Feedback from someone with more RDF expertize would be nice. CCing Lucas.

Anyway, I think that is the issue this patch got stuck on.

Patch-set: 6
Label: Code-Review=-1