Update patch set 1
Patch Set 1:

Sorry, perhaps I was not clear. The new version of eslint is stricter in terms of enforcing indentation rules, this means when it comes to:

 expression
   .method()

you have to choose a style and stick with it. As multiple styles are already present in this codebase, that means something needs to give. I went with the upstream default in wikimedia config (which is the same as eslint:recommended, to use indent with no extra options).

Patch-set: 1