Update patch set 1
Patch Set 1:

Thanks Ed for bringing this up!
And, yeah, this seems like a tricky one actually.
According to the Wikibase code style https://www.mediawiki.org/wiki/Wikibase/Coding_conventions#Indentation, it seems tempting to use the "MemberExpression: 0" option of the indent rule.
BUT: it is going to complain (a lot) for things like

var $( '<div/>' )
	.addClass( 'some-class' )
	.append( $( '<div/>' ).addClass( 'some-other-class' ) );

or

return $( '<div/>' )
	.addClass( 'some-class' )
	.append( $( '<div/>' ).addClass( 'some-other-class' ) );

which are actually the only correct way according to Wikibase style.

I actually thing the way of treating those two "special" cases is  nicer than not indenting. But as eslint's indent rule does not allow for exception I think the only options are:

1. Use "memberExcetion: off" to ignore this check - good as the tool would not complain about the "right" code but bad as would also not complain about the "wrong" code.

2. Change WB style so it always wants no indentation even on variable declaration on return statement

3. Change WB to adhere to more general/MW standard, i.e. indent member expression with a single tab.

I reckon that options 2 and 3 might be controversial, so maybe it might make sense to go with 1 for now, and have us (Wikibase folks) figure out whether we want to change the style, or try to teach the tool handle it?

Patch-set: 1